+++ src/main/java/com/microsoft/azure/functions/internal/spi/middleware/MiddlewareChain.java
@@ -3,18 +3,17 @@
  * Licensed under the MIT License. See License.txt in the project root for
  * license information.
  */
-package com.microsoft.azure.functions.middleware;
-
-import com.microsoft.azure.functions.internal.MiddlewareContext;
+package com.microsoft.azure.functions.internal.spi.middleware;
 
 /**
- * The function middleware chain.
+ * The middleware chain.
  *
- * @since 1.1.0
+ * <p>This class is internal and is hence not for public use at this time. Its APIs are unstable and can change
+ * at any time.
  */
-public interface FunctionMiddlewareChain {
+public interface MiddlewareChain {
     /**
-     * Invokes next middleware, usually used at the end of middleware to invoke next middleware in the middleware chain.
+     * Invokes next middleware in the chain.
      * @param context execution context that pass along middleware chain
      * @throws Exception any exception that happen along middleware chain
      */
